bot/requirements.txt
pyTelegramBotAPI==4.25.0
supabase==2.11.0
python-dotenv==1.0.1
